+++ lib/flutter_gutter.dart
@@ -1,3 +1,4 @@
 export 'src/breakpoints_utils.dart';
 export 'src/gap.dart';
 export 'src/gutter.dart';
+export 'src/gutter_type.dart';