+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 
 - Add optional size parameter
 - Introduce constructors on Gutter for types like small, large, etc
+- Add Gutter.expand
 - Add extension to add Gutter on Lists
 - Migrate to flutter_adaptive_scaffold to remove deprecated package
+- Add AdaptiveGutter that switches Gutter size based on the current breakpoint
 
 ## [1.0.3] - 07/27/2024